Simplify networking options
Before we were exposing methods to set custom client or service objects. This makes it more complicated to enforce default config on these objects. Since we only need a way to add a mock interceptor to the httpClient for testing, we're removing the custom setters and adding a method to enable a mock mode.