Force passing in content type rather than guessing
This has been an issue for awhile. Originally we were using Tika but
that doesn't work on Android. The Java library methods are utterly
useless and the alternative using URLConnection varies too much
per platform. It seems to work decently on Android but not in the JDK.
This will have to be sorted out but for now a user can just pass
it in manually or use a library that's appropriate for their platform
to guess it.